+++ examples/journald.py
@@ -4,7 +4,7 @@ Write some logs to journald.
 
 from __future__ import print_function
 
-from eliot import Message, start_action, add_destinations
+from eliot import log_message, start_action, add_destinations
 from eliot.journald import JournaldDestination
 
 add_destinations(JournaldDestination())
@@ -15,5 +15,5 @@ def divide(a, b):
         return a / b
 
 print(divide(10, 2))
-Message.log(message_type="inbetween")
+log_message(message_type="inbetween")
 print(divide(10, 0))